+++ CHANGELOG.md
@@ -337,3 +337,4 @@
 - [2026-04-12 12:28] feat(ui): implement shimmer loading skeleton in ui
 - [2026-04-12 12:51] test(storage): add unit tests for storage business logic
 - [2026-04-12 13:17] style(cache): harmonize light and dark theme palette for cache
+- [2026-04-12 13:19] feat(cache): implement shimmer loading skeleton in cache